Move UI source regression checks into separate test
